Contrôleur Authentification pointant sur le métier Utilisateur
