--- App/Components/Authentification/Dao.php
@@ -1,40 +1,0 @@
-<?php
-namespace App\Components\Authentification;
-
-/**
- * {@inheritDoc}
- *
- * @author Prytoegrian <prytoegrian@protonmail.com>
- * @author Wouldsmina
- *
- * @since 0.2
- *
- * Ne devrait être contacté que par Planning\Repository
- * Ne devrait contacter personne
- */
-class Dao extends \App\Libraries\ADao
-{
-    public function getById($id)
-    {
-    }
-
-    public function getList(array $a)
-    {
-    }
-
-    public function post(array $a)
-    {
-    }
-
-    public function put(array $a, $id)
-    {
-    }
-
-    public function delete($id)
-    {
-    }
-
-    public function getTableName()
-    {
-    }
-}
